% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artdeconv.R
\name{artdeconv}
\alias{artdeconv}
\title{Adaptive Regularized Tri-factor NMF Deconvolution}
\usage{
artdeconv(
  Y,
  Theta_0,
  m0 = nrow(Theta_0),
  k0,
  meds,
  ranges,
  alpha1,
  alpha2,
  beta,
  n_start = 10,
  parallel = TRUE,
  s_fixed = FALSE,
  ...
)
}
\arguments{
\item{Y}{The bulk matrix.}

\item{Theta_0}{The \eqn{m_0 \times K} reference signature matrix. The last \eqn{K_0} columns for cell types with unknown reference gene expression should be padded with 0s.}

\item{m0}{The number of signature genes for well characterized cell types in a tissue. Must satisfy \eqn{m_0 <= m}. The default value is \eqn{m} (\code{nrow(Theta_0)}).}

\item{k0}{The number of well characterized cell types in a tissue. Must satisfy \eqn{K_0 < K}.}

\item{meds}{A vector of length \eqn{K} of pre-specified medians of cell proportions.}

\item{ranges}{A vector of length \eqn{K} of pre-specified ranges of cell-type proportions (1/\code{ranges} is used as weights for regularization parameters).}

\item{alpha1}{The tuning parameter for regularizing the known part of Theta.}

\item{alpha2}{The tuning parameter for regularizing the unknown part of Theta.}

\item{beta}{The tuning parameter for regularizing P.}

\item{n_start}{An integer indicating the number of restarts of the algorithm; default value is 10.}

\item{parallel}{A logical value indicating whether or not to run \link{artdeconv} in parallel; default is \code{TRUE}.}

\item{s_fixed}{A logical value indicating whether the cell size matrix \eqn{S} is coerced to be the identity matrix; default is \code{FALSE}.}

\item{...}{Other parameters that can be passed to the function (see \link{artdeconv_single_solve_cpp}).}
}
\value{
A list with following items:
\itemize{
\item Y_hat: the estimated Y from the product of estimated factors;
\item Theta_hat: the estimated signature matrix from deconvolution;
\item s_hat: the estimated "cell size" vector from deconvolution;
\item P_hat the estimated Theta from deconvolution.
}
}
\description{
This function performs deconvolution of the bulk expression data using the ARTdeConv algorithm.
The user should input
}
\seealso{
\link{artdeconv_single_solve_cpp}
}
