% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artdeconv.R
\name{artdeconv_for_cv}
\alias{artdeconv_for_cv}
\title{The Bare Bone ARTDeConv Function For Cross-validations}
\usage{
artdeconv_for_cv(
  Y,
  Theta_0,
  m0,
  k0,
  meds,
  ranges,
  alpha1,
  alpha2,
  beta,
  n_start = 10,
  parallel = TRUE,
  ...
)
}
\description{
The Bare Bone ARTDeConv Function For Cross-validations
}
\seealso{
\link{artdeconv_single_solve_for_cv}
}
