% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_function.R
\name{artdeconv_obj_fun}
\alias{artdeconv_obj_fun}
\title{Objective function for the ARTDeConv algorithm}
\usage{
artdeconv_obj_fun(
  Y,
  Y_hat,
  Theta_hat,
  P_hat,
  m,
  n,
  k,
  m0,
  k0,
  Theta_0,
  meds,
  ranges,
  alpha_1,
  alpha_2,
  beta
)
}
\description{
Objective function for the ARTDeConv algorithm
}
