% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artdeconv_single_run.R
\name{artdeconv_single_solve}
\alias{artdeconv_single_solve}
\title{ARTdeConv function on a single run}
\usage{
artdeconv_single_solve(
  Y,
  Theta_0,
  Theta_it,
  s_it,
  P_it,
  m,
  n,
  k,
  m0,
  k0,
  meds,
  ranges,
  alpha1,
  alpha2,
  beta,
  max_iter = 1e+05,
  tol = 1e-05
)
}
\arguments{
\item{Y}{the bulk matrix}

\item{Theta_0}{\eqn{m_0 \times K_0} reference matrix for the fixed part of the signature matrix}

\item{Theta_it}{the initial value of the signature matrix}

\item{s_it}{the initial value of the "cell size" vector}

\item{P_it}{the initial value of the proportions matrix}

\item{m0}{the number of signature genes for well characterized cell types in a tissue}

\item{k0}{the number of well characterized cell types in a tissue}

\item{meds}{a vector of length \eqn{K} of pre-specified medians of cell proportions}

\item{ranges}{a vector of length \eqn{K} of pre-specified ranges of cell-type proportions (1/ranges as weights for regularization parameters)}

\item{alpha1}{the tuning parameter for regularizing the known part of Theta}

\item{alpha2}{the tuning parameter for regularizing the unknown part of Theta}

\item{beta}{the tuning parameter for regularizing P}

\item{max_iter}{the maximal number of iterations the algorithm shall perform; default value is \eqn{10^5}}

\item{tol}{tolerance for convergence indication; default value is \eqn{10^{-6}}}
}
\value{
A list with following items
\itemize{
\item{Y_hat the estimated Y from the product of estimated factors}
\item{Theta_hat the estimated signature matrix from deconvolution}
\item{s_hat the estimated "cell size" vector from deconvolution}
\item{P_hat the estimated Theta from deconvolution}
}
}
\description{
This is an artdeconv function on a single run.
}
