% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artdeconv_single_run.R
\name{artdeconv_single_solve_for_cv}
\alias{artdeconv_single_solve_for_cv}
\title{The Bare Bone Version of The Single Run For Cross Validation}
\usage{
artdeconv_single_solve_for_cv(
  Y,
  Theta_0,
  Theta_it,
  s_it,
  P_it,
  m,
  n,
  k,
  m0,
  k0,
  meds,
  ranges,
  alpha1,
  alpha2,
  beta,
  max_iter = 1e+05,
  tol = 1e-05
)
}
\description{
The Bare Bone Version of The Single Run For Cross Validation
}
