% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artdeconv_single_run.R
\name{artdeconv_single_solve_s_fixed}
\alias{artdeconv_single_solve_s_fixed}
\title{ARTdeConv function for a bi-factor model with fixed S}
\usage{
artdeconv_single_solve_s_fixed(
  Y,
  Theta_0,
  Theta_it,
  P_it,
  m,
  n,
  k,
  m0,
  k0,
  meds,
  ranges,
  alpha1,
  alpha2,
  beta,
  max_iter = 1e+05,
  tol = 1e-05
)
}
\description{
ARTdeConv function for a bi-factor model with fixed S
}
