% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_Delta}
\alias{get_Delta}
\title{Returning a Delta Matrix}
\usage{
get_Delta(k, k0, m, m0)
}
\arguments{
\item{k}{the total number of cell types (CTs) in the deconvolution}

\item{k0}{the total number of known CTs in the deconvolution (\eqn{\leq K})}

\item{m}{the total number of genes in the deconvolution}

\item{m0}{the total number of genes of the known CTs in the deconvolution (\eqn{\leq m})}
}
\value{
a \eqn{M \times K} binary matrix
}
\description{
Get a \eqn{M \times K} binary matrix (Delta matrix) where entries that are 1 correspond to
known expression values in the signature matrix and 0 to unknown values;
\strong{we always assume the missing cell types are in the last columns}
}
