% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_Delta_c_cpp}
\alias{get_Delta_c_cpp}
\title{Function for getting the complement of the Delta matrix}
\usage{
get_Delta_c_cpp(Delta)
}
\description{
Function for getting the complement of the Delta matrix
}
