% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inner_product_TL}
\alias{inner_product_TL}
\title{Function for Calculating the Frobenius Inner Products}
\usage{
inner_product_TL(A, B)
}
\arguments{
\item{A}{a matrix}

\item{B}{a matrix of the same size of \code{A}}
}
\value{
the Frobenius inner products of \code{A} and \code{B}
}
\description{
Return the Frobenius (trace) inner products of two matrices of
the same size; will return an error message if the size does
not match
}
