% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicative_updates.R
\name{mu_P}
\alias{mu_P}
\title{function to update the proportion matrix through multiplicative update}
\usage{
mu_P(Y, Theta, s, meds, P_old, m, n, beta, wt = 1)
}
\arguments{
\item{Y}{the bulk matrix}

\item{Theta}{the signature matrix (assumed fixed here)}

\item{s}{the diagonal "cell size" vector (assumed fixed here)}

\item{meds}{the vector of known medians}

\item{P_old}{the proportion matrix before the update}

\item{beta}{the tuning parameter for regularizing P}

\item{wt}{the inverse weights of cell types; must be a a vector of length k if using non-uniform weights or a scalar if using uniform weights}
}
\value{
the $K \times n$ proportion matrix after the current update
}
\description{
function to update the proportion matrix through multiplicative update
}
