% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicative_updates.R
\name{mu_s}
\alias{mu_s}
\title{function to update the size vector through multiplicative update}
\usage{
mu_s(Y, Theta, s_old, P)
}
\arguments{
\item{Y}{the bulk matrix}

\item{Theta}{the signature matrix (assumed fixed here)}

\item{s_old}{the "cell size" vector before the update}

\item{P}{the proportions matrix (assumed fixed here)}
}
\value{
the "cell size" vector of length $K$ after the update
}
\description{
function to update the size vector through multiplicative update
}
