% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicative_updates.R
\name{mu_Theta}
\alias{mu_Theta}
\title{function to update the signature matrix through multiplicative update}
\usage{
mu_Theta(Y, s, P, Theta_0, Delta, Delta_c, Theta_old, m, n, alpha1, alpha2)
}
\arguments{
\item{Y}{the bulk matrix}

\item{s}{the diagonal "cell size" vector (assumed fixed here)}

\item{P}{the proportions matrix (assumed fixed here)}

\item{Theta_0}{signature matrix for the fixed part of Theta}

\item{Delta}{indicator matrix for non-missing signature matrix expression values}

\item{Delta_c}{complement of the indicator matrix}

\item{Theta_old}{the signature matrix before the update}

\item{alpha1}{the tuning parameter for regularizing the known part of Theta}

\item{alpha2}{the tuning parameter for regularizing the unknown part of Theta}

\item{m0}{number of signature genes for well characterized cell types in a tissue}

\item{k0}{number of well characterized cell types in a tissue}
}
\value{
the $m \times K$ signature matrix after the update
}
\description{
function to update the signature matrix through multiplicative update
}
