% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{obj_fun_cpp}
\alias{obj_fun_cpp}
\title{ARTdeConv objective function}
\usage{
obj_fun_cpp(
  Y,
  Y_hat,
  Theta_hat,
  P_hat,
  m0,
  k0,
  Theta_0,
  Delta,
  Delta_c,
  meds,
  ranges,
  alpha_1,
  alpha_2,
  beta
)
}
\description{
ARTdeConv objective function
}
