% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{artdeconv_single_solve_s_fixed_cpp}
\alias{artdeconv_single_solve_s_fixed_cpp}
\title{Core ARTdeConv Function For One Set of Initial Values Assuming a Fixed S (in C++)}
\usage{
artdeconv_single_solve_s_fixed_cpp(
  Y,
  Theta_0,
  Theta_it,
  P_it,
  m0,
  k0,
  meds,
  ranges,
  alpha1,
  alpha2,
  beta,
  max_iter = 100000L,
  tol = 1e-05
)
}
\arguments{
\item{Y}{the bulk matrix.}

\item{Theta_0}{the partial reference matrix that we have prior knowledge about. The unknown part of}

\item{Theta_it}{the initial value of Theta.}

\item{P_it}{the initial value of P.}

\item{m0}{the number of gene features we have prior knowledge about in \code{Theta_0}.}

\item{k0}{the number of cell types whose cell-type level expression we have prior knowledge about in \code{Theta_0}.}

\item{meds}{the vector of median cell proportions in the population.}

\item{ranges}{the vector of the range interval lengths of proportions of different cell types.}

\item{alpha1}{the tuning parameter for regularizing the part of Theta with prior knowledge.}

\item{alpha2}{the tuning parameter for regularizing the part of Theta without prior knowledge.}

\item{beta}{the tuning parameter for regularizing P.}

\item{max_iter}{the maximal number of iterations this core function will run. The default is \code{1e5}.}

\item{tol}{the tolerance parameter for the convergence criterion of ARTdeConv. The default is \code{1e-5}.}
}
\description{
This is the core function that will run ARTdeConv once under a set of initial values written in C++, but with the legacy bi-factor assumption (that the scale matrix S is fixed),
and integrated in R through Rcpp and RcppArmadillo. It can technically increase the speed since S will not be updated,
but should be only used when the user is confident that S can be represented by the identity matrix. Users running this function should provide their own
initial values in the input. Otherwise, the required parameters are the same as the main ARTdeConv function
with restarts.
}
