% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_values.R
\name{get_initials}
\alias{get_initials}
\title{Function to Generate One Set of Initial Values For The ARTdeConv Algorithm}
\usage{
get_initials(Y_star, Theta_0, meds, m, n, k, m0, k0)
}
\arguments{
\item{Y_star}{bulk expression matrix}

\item{Theta_0}{the known part of \eqn{\mathbf{\Theta}}}

\item{meds}{a vector of length \eqn{K} of the pre-specified medians of cell type proportions}

\item{m}{the total number of genes in the deconvolution}

\item{k}{the total number of cell types (CTs) in the deconvolution}

\item{m0}{the total number of genes of the known CTs in the deconvolution (\eqn{\leq m})}

\item{k0}{the total number of known CTs in the deconvolution (\eqn{\leq K})}
}
\value{
A list of the initial values for \eqn{(\mathbf{\Theta}, \mathbf{s}, \mathbf{P})}.
}
\description{
Generate one set of \eqn{(\mathbf{\Theta}, \mathbf{s}, \mathbf{P})} based on the problem sizes and
known part of \eqn{\mathbf{\Theta}}. Details can be found in Liu et al.
}
